[PrizyPricer] injected ideal price formula, added filtered list view, changed type for Price.price (BigDecimal)

+++ PrizyPricer/grails-app/domain/prizypricer/Price.groovy
@@ -1,7 +1,7 @@
 package prizypricer
 
 class Price {
-	float price
+	BigDecimal price
 	String store
 	String notes
 	Date date
